+++ vulkano/src/pipeline/cache.rs
@@ -219,6 +219,8 @@ impl PipelineCache {
 unsafe impl VulkanObject for PipelineCache {
     type Object = vk::PipelineCache;
 
+    const TYPE: vk::DebugReportObjectTypeEXT = vk::DEBUG_REPORT_OBJECT_TYPE_PIPELINE_CACHE_EXT;
+
     #[inline]
     fn internal_object(&self) -> vk::PipelineCache {
         self.cache