+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix occasional truncation of glslang_validator when glsl-to-spirv is rebuilt
 - Fix linking against MoltenVK >= 0.19.0 
 - Added `AutoCommandBufferBuilder::copy_image`
+- Added `VulkanObject::TYPE` to look up the `DebugReportObjectTypeEXT` of an object
+- Added `Device::set_object_name` and `Device::set_object_name_raw`
 
 # Version 0.7.2 (2017-10-09)